refactor(e2e): fold validator harness into MultiNodeTestContext and split slashing/recovery
