+++ yarn-project/end-to-end/src/multi-node/block-production/first_slot.test.ts
