+++ yarn-project/end-to-end/src/multi-node/slashing/duplicate_proposal.test.ts
@@ -5,9 +5,19 @@ import { EpochNumber, SlotNumber } from '@aztec/foundation/branded-types';
 import { retryUntil } from '@aztec/foundation/retry';
 import { OffenseType } from '@aztec/slasher';
 
-import { advanceToEpochBeforeProposer, awaitCommitteeExists, awaitOffenseDetected } from '../../../e2e_p2p/shared.js';
-import { ValidatorRegistrationHarness } from '../../validator_registration_harness.js';
-import { AZTEC_SLOT_DURATION, aztecEpochDuration, baseHarnessOpts, slashingRoundSize } from './setup.js';
+import { advanceToEpochBeforeProposer, awaitCommitteeExists, awaitOffenseDetected } from '../../e2e_p2p/shared.js';
+import {
+  MultiNodeTestContext,
+  SLASHER_ENABLED_MULTI_VALIDATOR_OPTS,
+  buildMockGossipValidators,
+} from '../multi_node_test_context.js';
+import {
+  AZTEC_SLOT_DURATION,
+  NUM_VALIDATORS,
+  aztecEpochDuration,
+  baseSlashingOpts,
+  slashingRoundSize,
+} from './setup.js';
 
 /**
  * Test that slashing occurs when a validator sends duplicate proposals (equivocation).
@@ -20,23 +30,26 @@ import { AZTEC_SLOT_DURATION, aztecEpochDuration, baseHarnessOpts, slashingRound
  * 3. Since they have different coinbase addresses, their proposals will have different archives (different content)
  * 4. Other validators will detect the duplicate and emit a slash event
  *
- * Setup: MultiNodeTestContext via ValidatorRegistrationHarness on the in-memory mock-gossip bus (no real
- * libp2p). 4 validators, ethSlot=8s, aztecSlot=24s, epoch=2, proofSubEpochs=1024, minTxsPerBlock=0, inboxLag=2
- * (v5 always enforces the timetable).
+ * Setup: MultiNodeTestContext on the in-memory mock-gossip bus (no real libp2p). 4 validators, ethSlot=8s,
+ * aztecSlot=24s, epoch=2, proofSubEpochs=1024, minTxsPerBlock=0, inboxLag=2 (v5 always enforces the timetable).
  */
-describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
-  let harness: ValidatorRegistrationHarness;
+describe('multi-node/slashing/duplicate_proposal', () => {
+  let test: MultiNodeTestContext;
   let nodes: AztecNodeService[];
 
   beforeEach(async () => {
-    harness = await ValidatorRegistrationHarness.create({ ...baseHarnessOpts });
+    test = await MultiNodeTestContext.setup({
+      ...SLASHER_ENABLED_MULTI_VALIDATOR_OPTS,
+      ...baseSlashingOpts,
+      initialValidators: buildMockGossipValidators(NUM_VALIDATORS),
+    });
   });
 
   afterEach(async () => {
-    await harness.teardown();
+    await test.teardown();
   });
 
-  const cheatCodes = () => harness.context.context.cheatCodes;
+  const cheatCodes = () => test.context.cheatCodes;
 
   const debugRollup = async () => {
     await cheatCodes().rollup.debugRollup();
@@ -48,27 +61,27 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
   // observed by whichever node processed both proposals before the slot closed) and asserts the offense
   // is attributed to the shared key's address.
   it('slashes validator who sends duplicate proposals', async () => {
-    const { rollup } = await harness.getContracts();
+    const { rollup } = await test.getSlashingContracts();
 
     // Jump forward to an epoch in the future such that the validator set is not empty
     await cheatCodes().rollup.advanceToEpoch(EpochNumber(4));
     await debugRollup();
 
-    harness.logger.warn('Creating nodes');
+    test.logger.warn('Creating nodes');
 
     // Use validator index 0 for the "malicious" validator key
     const maliciousValidatorIndex = 0;
-    const maliciousValidatorAddress = harness.addressAt(maliciousValidatorIndex);
+    const maliciousValidatorAddress = test.addressAt(maliciousValidatorIndex);
 
-    harness.logger.warn(`Malicious proposer address: ${maliciousValidatorAddress.toString()}`);
+    test.logger.warn(`Malicious proposer address: ${maliciousValidatorAddress.toString()}`);
 
     // Create two nodes with the SAME validator key but DIFFERENT coinbase addresses
     // This will cause them to create proposals with different content for the same slot
     const coinbase1 = EthAddress.random();
     const coinbase2 = EthAddress.random();
 
-    harness.logger.warn(`Creating malicious node 1 with coinbase ${coinbase1.toString()}`);
-    const maliciousNode1 = await harness.createValidatorNode(maliciousValidatorIndex, {
+    test.logger.warn(`Creating malicious node 1 with coinbase ${coinbase1.toString()}`);
+    const maliciousNode1 = await test.createValidatorNodeAt(maliciousValidatorIndex, {
       coinbase: coinbase1,
       broadcastEquivocatedProposals: true,
       dontStartSequencer: true,
@@ -77,8 +90,8 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
       skipPushProposedBlocksToArchiver: true,
     });
 
-    harness.logger.warn(`Creating malicious node 2 with coinbase ${coinbase2.toString()}`);
-    const maliciousNode2 = await harness.createValidatorNode(maliciousValidatorIndex, {
+    test.logger.warn(`Creating malicious node 2 with coinbase ${coinbase2.toString()}`);
+    const maliciousNode2 = await test.createValidatorNodeAt(maliciousValidatorIndex, {
       coinbase: coinbase2,
       broadcastEquivocatedProposals: true,
       dontStartSequencer: true,
@@ -88,13 +101,13 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
     });
 
     // Create honest nodes with unique validator keys (indices 1 and 2)
-    harness.logger.warn('Creating honest nodes');
-    const honestNode1 = await harness.createValidatorNode(1, { dontStartSequencer: true });
-    const honestNode2 = await harness.createValidatorNode(2, { dontStartSequencer: true });
+    test.logger.warn('Creating honest nodes');
+    const honestNode1 = await test.createValidatorNodeAt(1, { dontStartSequencer: true });
+    const honestNode2 = await test.createValidatorNodeAt(2, { dontStartSequencer: true });
 
     nodes = [maliciousNode1, maliciousNode2, honestNode1, honestNode2];
 
-    await awaitCommitteeExists({ rollup, logger: harness.logger });
+    await awaitCommitteeExists({ rollup, logger: test.logger });
 
     // Find an epoch where the malicious proposer is selected, stopping one epoch before
     // so we have time to start sequencers before the target epoch arrives
@@ -103,11 +116,11 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
       epochCache,
       cheatCodes: cheatCodes().rollup,
       targetProposer: maliciousValidatorAddress,
-      logger: harness.logger,
+      logger: test.logger,
     });
 
     // Start all sequencers while still one epoch before the target
-    harness.logger.warn('Starting all sequencers');
+    test.logger.warn('Starting all sequencers');
     await Promise.all(nodes.map(n => n.getSequencer()!.start()));
 
     // Now warp to one slot before the target epoch — sequencers are already running.
@@ -116,7 +129,7 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
     // serialize past the slot boundary, after which honest receivers reject them as late. The helper
     // picks a target slot at least one slot into the epoch, so warping here leaves a full warm-up
     // slot before the build begins rather than starting it at the exact instant of the warp.
-    harness.logger.warn(`Advancing to one slot before target epoch ${targetEpoch} (target slot ${targetSlot})`);
+    test.logger.warn(`Advancing to one slot before target epoch ${targetEpoch} (target slot ${targetSlot})`);
     await cheatCodes().rollup.advanceToEpoch(targetEpoch, { offset: -AZTEC_SLOT_DURATION });
 
     // Wait for offense to be detected. Under proposer pipelining, checkpoint proposals are broadcast
@@ -126,10 +139,10 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
     // process both proposals while still in the build slot (often the other malicious node, since
     // they receive each other's broadcasts immediately). We therefore collect offenses from every
     // node in the network and assert that at least one of them recorded the duplicate proposal.
-    harness.logger.warn('Waiting for duplicate proposal offense to be detected...');
+    test.logger.warn('Waiting for duplicate proposal offense to be detected...');
     await awaitOffenseDetected({
       epochDuration: aztecEpochDuration,
-      logger: harness.logger,
+      logger: test.logger,
       nodeAdmin: honestNode1,
       slashingRoundSize,
       waitUntilOffenseCount: 1,
@@ -151,7 +164,7 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
       AZTEC_SLOT_DURATION * 4,
     );
 
-    harness.logger.warn(`Collected duplicate proposal offenses`, { proposalOffenses });
+    test.logger.warn(`Collected duplicate proposal offenses`, { proposalOffenses });
     expect(proposalOffenses.length).toBeGreaterThan(0);
     for (const offense of proposalOffenses) {
       expect(offense.validator.toString()).toEqual(maliciousValidatorAddress.toString());
@@ -161,10 +174,10 @@ describe('multi-node/slashing/equivocation/duplicate_proposal', () => {
     for (const offense of proposalOffenses) {
       const offenseSlot = SlotNumber(Number(offense.epochOrSlot));
       const proposerForSlot = await epochCache.getProposerAttesterAddressInSlot(offenseSlot);
-      harness.logger.info(`Offense slot ${offenseSlot}: proposer is ${proposerForSlot?.toString()}`);
+      test.logger.info(`Offense slot ${offenseSlot}: proposer is ${proposerForSlot?.toString()}`);
       expect(proposerForSlot?.toString()).toEqual(maliciousValidatorAddress.toString());
     }
 
-    harness.logger.warn('Duplicate proposal offense correctly detected and recorded');
+    test.logger.warn('Duplicate proposal offense correctly detected and recorded');
   });
 });
