+++ yarn-project/end-to-end/src/single-node/proving/upload_failed_proof.test.ts
