+++ yarn-project/end-to-end/src/single-node/proving/world_state_pruning.test.ts
