+++ yarn-project/end-to-end/src/single-node/recovery/manual_rollback.test.ts
