+++ yarn-project/end-to-end/src/single-node/recovery/prune_when_cannot_build.test.ts
